+++ src/helper/helpers.ts
@@ -1,21 +1,24 @@
 import OBR, { buildText } from "@owlbear-rodeo/sdk";
+import { textMetadata } from "./variables.ts";
 
 export const createText = async (text: string, id: string) => {
     const items = await OBR.scene.items.getItems([id]);
+    const width = 300;
+    const height = 20;
 
     if (items.length > 0) {
         const item = items[0];
         const position = {
-            x: item.position.x - 150,
-            y: item.position.y + 70,
+            x: item.position.x - width / 2,
+            y: item.position.y + 50 + height,
         };
 
-        return buildText()
+        const textItem = buildText()
             .text({
                 type: "PLAIN",
                 plainText: text,
-                height: 20,
-                width: 300,
+                height: height,
+                width: width,
                 style: {
                     fillColor: "#ffffff",
                     fillOpacity: 1,
@@ -36,7 +39,18 @@ export const createText = async (text: string, id: string) => {
             .attachedTo(id as string)
             .layer("TEXT")
             .locked(true)
-            .visible(true)
             .build();
+
+        textItem.metadata[textMetadata] = { isHpText: true };
+        return textItem;
     }
+};
+
+export const getAttachedTextItem = async (id: string) => {
+    const localItems = await OBR.scene.local.getItems();
+    const attachments = localItems.filter((item) => item.attachedTo === id);
+
+    return attachments.filter((attachment) => {
+        return attachment.layer === "TEXT" && textMetadata in attachment.metadata;
+    });
 };