+++ src/helper/types.ts
@@ -2,9 +2,10 @@ export type HpTrackerMetadata = {
     name: string;
     hp: number;
     maxHp: number;
+    armorClass: number;
     hpTrackerActive: boolean;
     canPlayersSee: boolean;
-    hpOnMap: string;
+    hpOnMap: boolean;
 };
 
 export type HpTextMetadata = {