+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 # 1.0.1
 
 - fix error when extension finishes loading before scene: https://github.com/owlbear-rodeo/extensions/pull/23#issuecomment-1629997777
+
+# 1.0.2
+
+- fix an error where HP is not refreshed in Main Window when changed in Popover
+- fixed invisible state by moving the text to local scene
+- handle move and visible state for local scene items
+- refactoring and minor optical improvements